Changed branches.sh to not return a commitid.
It was harmful anyway: git show-ref for tags returns the objectid of the
*tag*, not of its commit object.  You can check out using the tag object,
but then your checkout has a totally different object from the commitid that
revlist returns!  So just avoid the confusion by throwing away the objectids
from branches.sh altogether.

It turns out we were only using them in one place, and even that wasn't such
an *important* place, so it's fine.