Log parser: detect WvTest-style unit test failures.
These are lines of the form "! whatever   RESULT", where RESULT==ok means
success, and anything else is a test failure.  When we see such a line, we
report it in the test results, not as a fail, but as a warning.

Also add "<br>" lines where appropriate in the RSS feed.