Correctly report Warnings vs. Errors vs. Failures.
The previous logic would misreport the sort of problem depending on the
script's exit code.  Now we always report Warnings, Errors, and Failures
separately when we find them, *and* force an error if the exit code was
nonzero *and* there were no test failures (since a build error is the only
reason such a thing should happen).

The result is we can now see at a glance when a build fails because of unit
tests vs. build failures.