Add and use 'runtee' to avoid hangs due to long-running subprocesses.
If build.sh ended up launching a daemon or other program that ran in the
background, the autobuild cycle would hang, because that daemon might keep
its stdout open, and thus the "tee" it was feeding into would never exit.

Of course, daemons are supposed to close their stdout/stderr when forking
into the background, for exactly this reason.  But since we're an
autobuilder, we have no guarantee that the daemons we end up launching are
actually well-written and bug free, so we can't rely on this working.

Instead, runtee exits as soon as its subprogram dies *and* the input pipe
is empty.  It doesn't wait for the input pipe to *close*, however.