next-rev.sh: algorithm error would end bisection too early.
Technically, this was correct because we were searching for the "first
failing commit" (ie. the one that caused a transition from pass to
fail). However, in case there are multiple transitions because of a failing
commit in the middle, we should keep testing even after we find the
first failure.